--- client/server/src/router/index.js
@@ -1,7 +1,0 @@
-const express = require('express');
-const router = express.Router();
-const dictionaryRouter = require('./dictionary');
-
-router.use('/dictionary', dictionaryRouter);
-
-module.exports = router;
