+++ server/package.json
@@ -15,8 +15,8 @@
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node app.js",
-    "dev": "nodemon npm run start"
+    "start": "node ./src/app.js",
+    "dev": "nodemon node ./src/app.js"
   },
   "keywords": [],
   "author": "",
