+++ server/src/config/index.js
@@ -3,7 +3,7 @@ const dotenv = require('dotenv');
 dotenv.config();
 
 const dbConfig = {
-  host: process.env.DB_HOST,
+  // host: process.env.DB_HOST,
   database: process.env.DB_DATABASE,
   port: process.env.DB_PORT,
   username: process.env.DB_USERNAME,
