+++ server/src/middleware/user/index.js
@@ -40,97 +40,3 @@ const validUser = async (req, res, next) => {
 module.exports = {
   validUser,
 };
-router.post('/', async (req, res) => {
-  if (
-    !req.body.id ||
-    !req.body.email ||
-    !req.body.username ||
-    !req.body.password ||
-    !req.body.creatorGroup ||
-    !req.body.userGroup
-  )
-    return res.status(400).send({
-      message: 'infos are required',
-    });
-
-  try {
-    const existingUser = await User.findOne({
-      username: req.body.id,
-    });
-    if (existingUser)
-      return res.status(403).send({
-        message: 'id already exists',
-      });
-    const user = new User({
-      username: req.body.username,
-      password: req.body.password,
-      id: req.body.id,
-      email: req.body.email,
-      creatorGroup: req.body.creatorGroup,
-      userGroup: req.body.userGroup,
-    });
-    await user.save();
-
-    req.session.userID = user._id;
-
-    return res.send({
-      user: user,
-    });
-  } catch (error) {
-    console.log(error);
-    return res.sendStatus(500);
-  }
-});
-
-router.post('/login', async (req, res) => {
-  if (!req.body.id || !req.body.password) return res.sendStatus(400);
-
-  try {
-    const user = await User.findOne({
-      id: req.body.id,
-    });
-
-    if (!user)
-      return res.status(403).send({
-        message: 'id or password is wrong',
-      });
-
-    if (!(await user.comparePassword(req.body.passwrod)))
-      return res.status(403).send({
-        message: 'id or password is wrong',
-      });
-
-    req.session.userID = user._id;
-
-    return res.send({
-      user: user,
-    });
-  } catch (error) {
-    console.log(error);
-    return res.sendStatus(500);
-  }
-});
-
-router.get('/', validUser, async (req, res) => {
-  try {
-    res.send({
-      user: req.user,
-    });
-  } catch (error) {
-    console.log(error);
-    return res.sendStatus(500);
-  }
-});
-
-//로그아웃
-router.delete('/', validUser, async (req, res) => {
-  try {
-    req.session = null;
-    res.sendStatus(200);
-  } catch (error) {
-    console.log(error);
-    return res.sendStatus(500);
-  }
-});
-
-module.exports = router;
